packages/simulation-core/pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "movici-simulation-core"
dynamic = ["version"]
description = "Core package for running Movici geospatial temporal simulations"
readme = "README.rst"
requires-python = ">=3.10,<3.14"
license = { text = "Movici Public License" }
authors = [{ name = "NGinfra Movici", email = "movici@nginfra.nl" }]
maintainers = [{ name = "NGinfra - Movici", email = "movici@nginfra.nl" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: Free for non-commercial use",
  "License :: Other/Proprietary License",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
  "numpy",
  "numba>=0.55",
  "python-dateutil>=2.8.0",
  "pandas>=1.2.0",
  "pydantic>=1.8.2",
  "pydantic_settings",
  "orjson>=3",
  "pyzmq",
  "msgpack",
  "click",
  "geopandas",
  "jsonschema",
  "pyproj",
  "netCDF4",
  "Fiona>=1.9",
  "pyparsing",
]

[project.optional-dependencies]
models = [
  "shapely>=1.7.1",
  "aequilibrae>=1",
  "pyproj>=3.0.1",
  "movici-geo-query>=1.1.1",
  "scipy",
  "wntr>=1.0.0",
  # setuptools<81 supplies pkg_resources, required for wntr on older python versions
  "setuptools<81; python_version < '3.12'",
]
sqlite = ["sqlalchemy>=1.4,<3.0"]
all = [
  "shapely>=1.7.1",
  "aequilibrae>=1",
  "pyproj>=3.0.1",
  "movici-geo-query>=1.1.1",
  "scipy",
  "wntr>=1.0.0",
  "sqlalchemy>=1.4,<3.0",
]

[project.urls]
Homepage = "https://github.com/nginfra/movici-simulation-core"
Documentation = "https://docs.movici.nl/en/latest/index.html"
Repository = "https://github.com/nginfra/movici-simulation-core"
Issues = "https://github.com/nginfra/movici-simulation-core/issues"

[project.scripts]
movici-simulation = "movici_simulation_core.cli:main"

[project.entry-points."movici.plugins"]
orchestrator = "movici_simulation_core.services.orchestrator.service:Orchestrator"
update_data_service = "movici_simulation_core.services.update_data.service:UpdateDataService"
init_data_service = "movici_simulation_core.services.init_data.service:InitDataService"
global_attributes = "movici_simulation_core.attributes:GlobalAttributes"
common_attributes = "movici_simulation_core.models.common.attributes:CommonAttributes"
area_aggregation = "movici_simulation_core.models.area_aggregation.model:Model"
corridor = "movici_simulation_core.models.corridor.model:Model"
csv_player = "movici_simulation_core.models.csv_player.csv_player:CSVPlayer"
data_collector = "movici_simulation_core.models.data_collector.data_collector:DataCollector"
evacuation_point_resolution = "movici_simulation_core.models.evacuation_point_resolution:EvacuatonPointResolution"
generalized_journey_time = "movici_simulation_core.models.generalized_journey_time.gjt_model:GJTModel"
netcdf_player = "movici_simulation_core.models.netcdf_player.netcdf_player:NetCDFPlayer"
operational_status = "movici_simulation_core.models:OperationalStatusModel"
opportunities = "movici_simulation_core.models.opportunities.model:Model"
overlap_status = "movici_simulation_core.models.overlap_status.model:Model"
shortest_path = "movici_simulation_core.models.shortest_path.model:ShortestPathModel"
tape_player = "movici_simulation_core.models.tape_player.model:Model"
time_window_status = "movici_simulation_core.models.time_window_status.model:Model"
traffic_assignment_calculation = "movici_simulation_core.models.traffic_assignment_calculation.model:Model"
traffic_demand_calculation = "movici_simulation_core.models.traffic_demand_calculation.model:TrafficDemandCalculation"
traffic_kpi = "movici_simulation_core.models.traffic_kpi.model:Model"
udf = "movici_simulation_core.models.udf_model.udf_model:UDFModel"
unit_conversions = "movici_simulation_core.models.unit_conversions.model:Model"
rules = "movici_simulation_core.models.rules.model:Model"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv.build-backend]
data = { scripts = "bin" }

[tool.uv-dynamic-versioning]
fallback-version = "0.0.0"
